api/requirements.txt
# argo version with 3 added to major version (argo 3.4.8 == argo-workflows 6.4.8) https://github.com/argoproj/argo-workflows/tree/master/sdks/python#versioning
PyYAML==6.0.1
requests==2.32.3
fastapi==0.111.1
uvicorn[standard]==0.30.1
httpx-oauth==0.15.0
pydantic-settings==2.3.4
google-cloud-logging==3.10.0
pyproj==3.6.1
wget==3.2
pandas==2.2.2